finalizando con la 2 y 3 funcionalidad de esta subasta donde validar que 2 subasta no entren el puja el mismo producto ademas crear el endpoint  para apuntarse a la subasta

+++ Subasta-service/Subasta-service/src/main/java/com/Subasta_Online/Subasta_service/Model/DTOiniciarSubasta.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.Duration;
 import java.time.LocalDateTime;
 
@@ -17,7 +18,7 @@ public class DTOiniciarSubasta {
     private String nombre;
     private Categoria categoria;
     private String descripcion;
-    private float precioInicial;
+    private BigDecimal precioInicial;
     private LocalDateTime horaInicio;
     private EstadoProducto estadoProducto;
     private Duration duracionSubasta;
